+++ target/debug/examples/01_get_device-4687dd2594faf22c.d
@@ -1,9 +1,0 @@
-/Users/perjohansson/dev/apple-metal-rs/target/debug/examples/01_get_device-4687dd2594faf22c.d: examples/01_get_device.rs Cargo.toml
-
-/Users/perjohansson/dev/apple-metal-rs/target/debug/examples/lib01_get_device-4687dd2594faf22c.rmeta: examples/01_get_device.rs Cargo.toml
-
-examples/01_get_device.rs:
-Cargo.toml:
-
-# env-dep:CLIPPY_ARGS=-D__CLIPPY_HACKERY__warnings__CLIPPY_HACKERY__
-# env-dep:CLIPPY_CONF_DIR