+++ target/debug/examples/01_get_device-7f934ed4e3e34db3.d
@@ -1,5 +1,0 @@
-/Users/perjohansson/dev/apple-metal-rs/target/debug/examples/01_get_device-7f934ed4e3e34db3.d: examples/01_get_device.rs
-
-/Users/perjohansson/dev/apple-metal-rs/target/debug/examples/01_get_device-7f934ed4e3e34db3: examples/01_get_device.rs
-
-examples/01_get_device.rs: