+++ target/debug/examples/02_caps_buffer_texture-97c5a31ecd45d6b1.d
@@ -1,5 +1,0 @@
-/Users/perjohansson/dev/apple-metal-rs/target/debug/examples/02_caps_buffer_texture-97c5a31ecd45d6b1.d: examples/02_caps_buffer_texture.rs
-
-/Users/perjohansson/dev/apple-metal-rs/target/debug/examples/02_caps_buffer_texture-97c5a31ecd45d6b1: examples/02_caps_buffer_texture.rs
-
-examples/02_caps_buffer_texture.rs: