+++ target/debug/examples/02_caps_buffer_texture-ee107887edfb0914.d
@@ -1,9 +1,0 @@
-/Users/perjohansson/dev/apple-metal-rs/target/debug/examples/02_caps_buffer_texture-ee107887edfb0914.d: examples/02_caps_buffer_texture.rs Cargo.toml
-
-/Users/perjohansson/dev/apple-metal-rs/target/debug/examples/lib02_caps_buffer_texture-ee107887edfb0914.rmeta: examples/02_caps_buffer_texture.rs Cargo.toml
-
-examples/02_caps_buffer_texture.rs:
-Cargo.toml:
-
-# env-dep:CLIPPY_ARGS=-D__CLIPPY_HACKERY__warnings__CLIPPY_HACKERY__
-# env-dep:CLIPPY_CONF_DIR