+++ target/debug/examples/03_command_buffer_blit-d7afc697c76c3bee.d
@@ -1,5 +1,0 @@
-/Users/perjohansson/dev/apple-metal-rs/target/debug/examples/03_command_buffer_blit-d7afc697c76c3bee.d: examples/03_command_buffer_blit.rs
-
-/Users/perjohansson/dev/apple-metal-rs/target/debug/examples/03_command_buffer_blit-d7afc697c76c3bee: examples/03_command_buffer_blit.rs
-
-examples/03_command_buffer_blit.rs: